Fundamental Kotlin - Abstraction.
