--- CodeExamples/Fundamental-Kotlin/src/main/java/net/milosvasic/fundamental/kotlin/classes/Abstraction.kt
@@ -1,22 +1,0 @@
-package net.milosvasic.fundamental.kotlin.classes
-
-/**
- * Created by milosvasic on 6/8/16.
- */
-abstract class Animal {
-    abstract fun walk()
-
-    fun run(){
-        println("run")
-    }
-}
-
-open class Cat : Animal() {
-    override fun walk() {
-        println("walk")
-    }
-}
-
-abstract class BigCat : Cat() {
-    override abstract fun walk()
-}
